+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Bot now use only necessary gateway intents (discord API)
 - Added proper logger Plugin
 - Added embed showing player status when using =r command
+- Added basic stats display
 
 # v3.1:
 - Fixed some issues with sub command